requirements.txt
asgiref==3.7.2
Django==4.2.4
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.1
phonenumbers==8.13.18
psycopg2-binary==2.9.7
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
uritemplate==4.1.1
